+++ src/hooks/useShortState.ts
@@ -1,16 +1,30 @@
 import { useState, useRef } from 'react';
 import { useOnMount, useConst } from './hooks';
 
+const DEFAULT_DURATION = 0;
+
 // TODO: Add function signtures allowing either zero args, or 2.
-function useShortState<S>(initialState?: S | (() => S), duration?: number) {
+
+/** State variable that always snaps back to `undefined` after `duration` milliseconds. */
+const useShortState = <S>(
+	/** Initial temporary state that then gets reverted back
+	 * to `undefined` after `duration` milliseconds
+	 * */
+	initialState?: S | (() => S),
+
+	/** Default duration, can be overridden on a case-by-case basis
+	 * by passing a custom duration to the `setState` function
+	 * */
+	defaultDuration = DEFAULT_DURATION
+) => {
 	const [state, _setState] = useState<S | undefined>(initialState);
-	const timeout = useRef<NodeJS.Timeout | null>();
+	const timeout = useRef<ReturnType<typeof setTimeout> | null>();
 
 	const cancelTimeout = () => {
 		timeout.current && clearTimeout(timeout.current);
 	};
 
-	const setState = useConst((newState: S | (() => S), duration: number) => {
+	const setState = useConst((newState: S | (() => S), duration = defaultDuration) => {
 		_setState(newState);
 		cancelTimeout();
 		timeout.current = setTimeout(() => {
@@ -21,12 +35,12 @@ function useShortState<S>(initialState?: S | (() => S), duration?: number) {
 
 	useOnMount(() => {
 		if (initialState !== undefined) {
-			setState(initialState, duration || 0);
+			setState(initialState, defaultDuration);
 		}
 		return cancelTimeout;
 	});
 
 	return [state, setState] as const;
-}
+};
 
 export default useShortState;